Bundle calculations in one file
